Move server names functionality to dedicated package
